fix: improve azcopy.sh to check for existing AzCopy before downloading

+++ scripts/azcopy.sh
@@ -2,16 +2,23 @@
 # Download 64-bit version of AzCopy if on x86_64 architecture
 # https://aka.ms/downloadazcopy-v10-linux
 
-if [ "$(uname -m)" = "x86_64" ]; then
-    wget -O azcopy.tar.gz https://aka.ms/downloadazcopy-v10-linux
-fi
+# See if azcopy exists in this directory
+if [ -f 'azcopy' ]; then
+    echo "AzCopy already exists in this directory. Skipping download."
+else
+    echo "Downloading AzCopy..."
 
-# ARM64 architecture
-if [ "$(uname -m)" = "aarch64" ]; then
-    wget -O azcopy.tar.gz https://aka.ms/downloadazcopy-v10-linux-arm64
-fi
+    if [ "$(uname -m)" = "x86_64" ]; then
+        wget -O azcopy.tar.gz https://aka.ms/downloadazcopy-v10-linux
+    fi
 
-tar -xf azcopy.tar.gz --strip-components=1 --wildcards '*/azcopy'
+    # ARM64 architecture
+    if [ "$(uname -m)" = "aarch64" ]; then
+        wget -O azcopy.tar.gz https://aka.ms/downloadazcopy-v10-linux-arm64
+    fi
+
+    tar -xf azcopy.tar.gz --strip-components=1 --wildcards '*/azcopy'
+fi
 
 # First argument is the backup file, the second is supposed to be the URL + SAS token
 # set as an environment variable.
